move string utilities to its own package
